+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
 
+- Add OpenAPI version conversion to Postman Collection version (handy for usage in Microcks) (#577)
+- Bumped dependencies: openapi-to-postman 4.21.0
+
 ## v1.26.5 - (2024-03-08)
 
 - headersPresent: Add check only for required headers (#566)